Fix races in ozy send_command.
This code is still not good.  It should work as long as the responses
from whatever ozy is talking to are under the 64 byte buffer length
used for receive, because command send/receive are in lockstep.
However, depending on TCP PUSH is never good, and will bite us down
the road some time.