dspserver functionality changed in order to accomodate both remote and local transmissions. dspserver command line options Hermes related are changed as follows:
   --hpsdr (if using hpsdr hardware with no local mike and headphone)
   --hpsdrloc (if using hpsdr hardware with LOCAL mike and headphone)

in the first case the microphone audio is taken from QtRadio,
processed in dspserver and sent Hermes via the local hpsdr_server.

in the second case the audio stream arriving from QtRadio is ignored,
and the mic audio taken in account and processed is that coming from the hpsdr_server's I/Q stream.

Please note that, whilst not transmitting, the Hermes headphone output is enabled in both cases.